requirements.txt
django
pillow
django-cors-headers
djangorestframework
django-filter
markdown